Fix: Renames domain picture objects and deletes unused objects
