--- src/main/java/com/idealista/infrastructure/api/PublicAd.java
@@ -1,61 +1,0 @@
-package com.idealista.infrastructure.api;
-
-import java.util.List;
-
-public class PublicAd {
-
-    private Integer id;
-    private String typology;
-    private String description;
-    private List<String> pictureUrls;
-    private Integer houseSize;
-    private Integer gardenSize;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getTypology() {
-        return typology;
-    }
-
-    public void setTypology(String typology) {
-        this.typology = typology;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public List<String> getPictureUrls() {
-        return pictureUrls;
-    }
-
-    public void setPictureUrls(List<String> pictureUrls) {
-        this.pictureUrls = pictureUrls;
-    }
-
-    public Integer getHouseSize() {
-        return houseSize;
-    }
-
-    public void setHouseSize(Integer houseSize) {
-        this.houseSize = houseSize;
-    }
-
-    public Integer getGardenSize() {
-        return gardenSize;
-    }
-
-    public void setGardenSize(Integer gardenSize) {
-        this.gardenSize = gardenSize;
-    }
-}
