+++ src/main/java/com/idealista/infrastructure/api/ad/score/service/scorers/impl/DescriptionWordLengthAdScorer.java
