+++ src/main/java/com/idealista/infrastructure/api/ad/search/domain/AdPictureQuality.java
@@ -1,0 +1,7 @@
+package com.idealista.infrastructure.api.ad.search.domain;
+
+public enum PictureQuality {
+
+    HD, SD
+
+}
