--- src/main/java/com/idealista/infrastructure/persistence/PictureQuality.java
@@ -1,7 +1,0 @@
-package com.idealista.infrastructure.persistence;
-
-public enum PictureQuality {
-
-    HD, SD
-
-}
